add carusel, add query params - ins list <-> ins detail

+++ src/app/instructors/instructor-edit.component.ts
@@ -11,20 +11,22 @@ import { NgForm } from "@angular/forms";
   templateUrl: './instructor-edit.component.html'
 })
 export class InstructorEditComponent implements OnInit, OnDestroy {
-  startDate: Date;
+  // startDate: Date;
   // startDate: Date = new Date();
   // startDate: Date = new Date('Jan 1 2017');
-  minDate: Date = new Date('Oct 12 2016');
-  startTime: Date = new Date('Oct 12 2016 3:00 PM');
-  onOffSwitch: any = 'Off';
+  // minDate: Date = new Date('Oct 12 2016');
+  // startTime: Date = new Date('Oct 12 2016 3:00 PM');
+  // onOffSwitch: any = 'Off';
 
   subscriptioN: Subscription;
   isNewInstructor: boolean;
-  pageTitle: string = 'Instructor Edit';
+  pageTitle: string = '';
   // instructor: IInstructor = new Instructor();
   instructor$: Observable<IInstructor>;
 
   hasGearboxError: boolean = false;
+  defaultPhotoUrl: string = 'https://firebasestorage.googleapis.com/v0/b/instructor-dp-ua.appspot.com/o/instructors%2F8?alt=media&token=4337ceb0-a730-4ddc-8573-b6d06ed67887';
+  defaultCarPhotoUrl: string = '';
 
   constructor(
     private afDbService: AfDbService,
@@ -32,16 +34,22 @@ export class InstructorEditComponent implements OnInit, OnDestroy {
     private activatedRoute: ActivatedRoute) { }
 
   ngOnInit() {
-    let id: string = this.activatedRoute.snapshot.paramMap.get('id');
     // let id: string = this.activatedRoute.snapshot.params['id'];
-    this.pageTitle += `: ${id}`;
-    this.isNewInstructor = id === 'new';
 
-    if (!this.isNewInstructor) {
-      this.instructor$ = this.getInstructor(id);
-    } else {
-      this.instructor$ = Observable.of({}) as Observable<IInstructor>;
-    }
+    // Angular 4 introduced another ActivatedRoute interface called paramMap
+    // let id: string = this.activatedRoute.snapshot.paramMap.get('id');
+
+    this.activatedRoute.paramMap.subscribe(paramMap => {
+      let id: string = paramMap.get('id');
+      console.log(id);
+      this.isNewInstructor = id === 'new';
+      this.pageTitle = this.isNewInstructor ? 'Add Instructor' : `Edit Instructor: ${id}`;
+      if (!this.isNewInstructor) {
+        this.instructor$ = this.getInstructor(id);
+      } else {
+        this.instructor$ = Observable.of({}) as Observable<IInstructor>;
+      }
+    });
 
     // this.subscriptioN = this.instructor$
     //   .subscribe(
@@ -67,31 +75,41 @@ export class InstructorEditComponent implements OnInit, OnDestroy {
   }
 
   saveInstructor2in1(instructor: IInstructor) {
+    if (!instructor.hasOwnProperty('photoUrl')) instructor.photoUrl = this.defaultPhotoUrl;
+    if (!instructor.hasOwnProperty('carPhotoUrl')) instructor.carPhotoUrl = this.defaultCarPhotoUrl;
     const save = this.isNewInstructor
       ? this.afDbService.createInstructor(instructor)
       : this.afDbService.updateInstructor(instructor);
     save.then(_ => {
-      console.log('saveInstructor 2in1 navigate to /instructors');
+      alert(`Here will be MODAL. Now ALERT. ${instructor.instructorName} was ${(this.isNewInstructor) ? 'added' : 'updated'} instructor info.`);
+      // console.log('saveInstructor 2in1 navigate to /instructors');
       this.router.navigate(['/instructors']);
     });
   }
 
   deleteInstructor(instructor: IInstructor) {
-    this.afDbService.deleteInstructor(instructor)
+    if (confirm(`Here will be MODAL. Now CONFIRM. Delete instructor - ${instructor.instructorName}?`)) {
+      this.afDbService.deleteInstructor(instructor)
       .then(_ => {
-        console.log('deleteInstructor navigate to /instructors');
+        alert(`Here will be MODAL. Now ALERT.  ${instructor.instructorName} was deleted.`);
+        // console.log('deleteInstructor navigate to /instructors');
         this.router.navigate(['/instructors']);
       });
+    } else {
+      alert('Deleting was canceled!')
+    }
+
+
   }
 
   onBack() {
     this.router.navigate(['/instructors']);
   }
 
-  save(instructorForm: NgForm) {
+/*   save(instructorForm: NgForm) {
     console.log(instructorForm.form);
     console.log('Saved: ' + JSON.stringify(instructorForm.value));
-  }
+  } */
 
   validateGearbox(value) {
     console.log(`gearbox: this.instructor.gearbox , value: ${value}`);
